refactor(shape-plugin): split build-progress and config hooks modules
