+++ plugins/shape-plugin/src/ui/components/build-config/CacheManagementSection.tsx
@@ -16,7 +16,7 @@ import {
   Inventory2 as Inventory2Icon,
 } from '@mui/icons-material';
 import type { ShapeBuildConfig } from '../../../common/types/index.js';
-import type { FetchConfigSectionState } from './useFetchConfigSection.ts';
+import type { FetchConfigSectionState } from '../../hooks/useFetchConfigSection.ts';
 import { useTranslation } from '../../i18n.js';
 import { getBuildConfigHoverCardSx } from '@hierarchidb/ui-accordion-config';
