+++ plugins/shape-plugin/src/ui/components/build-progress/ShapeBuildProgressPanel/ShapeBuildProgressPanel.tsx
@@ -1,4 +1,4 @@
-import { type MouseEvent as ReactMouseEvent, type ReactNode, useCallback, useEffect, useId, useMemo, useRef, useState } from 'react';
+import { type MouseEvent as ReactMouseEvent, useCallback, useEffect, useMemo, useState } from 'react';
 import {
   Alert,
   Box,
@@ -7,692 +7,34 @@ import {
   DialogActions,
   DialogContent,
   DialogTitle,
-  IconButton,
   Popover,
-  Skeleton,
   Snackbar,
   Stack,
   Typography,
   Tooltip,
-  useTheme,
 } from '@mui/material';
 import ConstructionIcon from '@mui/icons-material/Construction';
-import ArrowCircleDownIcon from '@mui/icons-material/ArrowCircleDown';
-import ArrowCircleUpIcon from '@mui/icons-material/ArrowCircleUp';
 import DownloadingIcon from '@mui/icons-material/Downloading';
 import HourglassTopIcon from '@mui/icons-material/HourglassTop';
 import TimelapseIcon from '@mui/icons-material/Timelapse';
 import { type NodeId, toNodeType } from '@hierarchidb/core-types';
-import { BuildSessionProgressPanel, useBuildStageFilter } from '@hierarchidb/components';
+import { BuildSessionProgressPanel } from '@hierarchidb/components';
 import { BuildSessionLauncherPanel } from '@hierarchidb/ui-batch-progress';
 import { DownloadRetryControls, type DownloadRetryConfig, WorkerNumberConfigCard } from '@hierarchidb/ui-accordion-config';
 import { TreeTableSearchInput } from '@hierarchidb/ui-search-input';
-import { useAtomValue, useSetAtom } from 'jotai';
-import type { TaskItemWithMetadata } from './TaskItemCardListCard.tsx';
-import { TaskItemCardListCard, sortTransformTasks, sortVectorTileTasks } from './TaskItemCardListCard.tsx';
-import type { ShapeEntity } from '../../../common/types/ShapeEntity.ts';
-import type { ShapeProcessingConfig } from '../../../common/types/index.js';
+import type { TaskItemWithMetadata } from '../useTaskItemCardList.js';
+import type { ShapeEntity } from '../../../../common/types/ShapeEntity.js';
+import type { ShapeProcessingConfig } from '../../../../common/types/index.js';
 import {
   DEFAULT_BUILD_CONFIG,
   DEFAULT_PROCESSING_CONFIG,
   mergeBuildConfig,
   mergeProcessingConfig,
-} from '../../../common/types/index.js';
-import { isTaskSkipped } from '../../../common/utils/taskMessages.ts';
-import { useShapeBuildProgressPanel } from './useShapeBuildProgressPanel.ts';
-import { useShapeBuildCacheActions } from '../build-config/useShapeBuildCacheActions.ts';
-import type { TaskProgressSummary } from '../../atoms/shapeBuildProgressAtoms.js';
-import { taskScrollTargetAtom, taskViewportRangeAtom } from '../../atoms/shapeBuildProgressAtoms.js';
-import type { BuildStage } from '@hierarchidb/components/build-stage';
-
-const TaskProgressBar = ({
-  stages,
-  tasksByStage,
-  stageTotals,
-  buildStatus,
-  activeStageId,
-  resolveTaskTitle,
-}: {
-  stages: BuildStage[];
-  tasksByStage: Record<string, TaskItemWithMetadata[]>;
-  stageTotals?: TaskProgressSummary['stageTotals'];
-  buildStatus: TaskProgressSummary['buildStatus'];
-  activeStageId?: string | null;
-  resolveTaskTitle: (task: TaskItemWithMetadata) => string;
-}) => {
-  const theme = useTheme();
-  const filter = useBuildStageFilter();
-  const flowBandClipId = useId().replace(/:/g, '');
-  const setScrollTarget = useSetAtom(taskScrollTargetAtom);
-  const viewportRange = useAtomValue(taskViewportRangeAtom);
-  const waitingColor = theme.palette.grey[300];
-  const emptyStageColor = buildStatus === 'failed' ? theme.palette.error.main : theme.palette.grey[500];
-  const runningColor = theme.palette.info.main;
-  const failedColor = theme.palette.error.main;
-  const skippedColor = theme.palette.warning.main;
-  const viewportStageId = viewportRange?.stageId;
-  const viewportStartId = viewportRange?.startTaskId;
-  const viewportEndId = viewportRange?.endTaskId;
-  let viewportStartGlobal: number | null = null;
-  let viewportEndGlobal: number | null = null;
-  let viewportStartIndex: number | null = null;
-  let viewportEndIndex: number | null = null;
-  if (viewportStageId && viewportStartId && viewportEndId) {
-    const stageTasks = tasksByStage[viewportStageId] ?? [];
-    const orderedStageTasks = viewportStageId === 'vt'
-      ? sortVectorTileTasks(stageTasks)
-      : viewportStageId === 'transform'
-        ? sortTransformTasks(stageTasks)
-      : stageTasks;
-    const startIndex = orderedStageTasks.findIndex((task) => task.taskId === viewportStartId);
-    const endIndex = orderedStageTasks.findIndex((task) => task.taskId === viewportEndId);
-    if (startIndex >= 0 && endIndex >= 0) {
-      viewportStartIndex = Math.min(startIndex, endIndex);
-      viewportEndIndex = Math.max(startIndex, endIndex);
-    }
-  }
-  const successColor = theme.palette.success.main;
-  const pausedColor = theme.palette.warning.main;
-  const { segments, stageOffsets, stageCounts } = useMemo(() => {
-    const nextSegments: Array<{ fill: string; fillOpacity: number; stageId: string; taskId?: string; title: string; width: number }> = [];
-    const nextStageOffsets = new Map<string, number>();
-    const nextStageCounts = new Map<string, number>();
-    let totalCount = 0;
-    stages.forEach((stage) => {
-      const fallbackStageId = stage.id === 'transform'
-        && stages.length === 1
-        && (tasksByStage.transform?.length ?? 0) === 0
-        && (tasksByStage.fetch?.length ?? 0) > 0
-        ? 'fetch'
-        : stage.id;
-      const sourceStageId = fallbackStageId;
-      const stageTasks = tasksByStage[sourceStageId] ?? [];
-      nextStageOffsets.set(stage.id, totalCount);
-      const plannedStageTotal = Math.max(0, stageTotals?.[stage.id]?.total ?? 0);
-      const orderedTasks = stage.id === 'vt'
-        ? sortVectorTileTasks(stageTasks)
-        : stage.id === 'transform'
-          ? sortTransformTasks(stageTasks)
-        : stageTasks;
-      const expectedStageTotal = Math.max(orderedTasks.length, plannedStageTotal);
-      nextStageCounts.set(stage.id, expectedStageTotal);
-      orderedTasks.forEach((task) => {
-        const statusValue = (task.status ?? '').toString().toLowerCase();
-        let fill = waitingColor;
-        const isSkipped = isTaskSkipped(task.display, task.message);
-        if (isSkipped) {
-          fill = skippedColor;
-        } else if (statusValue === 'completed' || statusValue === 'recycled') {
-          fill = successColor;
-        } else if (statusValue === 'failed') {
-          fill = failedColor;
-        } else if (statusValue === 'running') {
-          fill = runningColor;
-        } else if (statusValue === 'paused') {
-          fill = pausedColor;
-        }
-        const isDimmed =
-          (isSkipped && !filter.skippedMode)
-          || (statusValue === 'failed' && !filter.failedMode)
-          || ((statusValue === 'completed' || statusValue === 'recycled') && !filter.completedMode);
-        const isExternalStage = sourceStageId !== stage.id;
-        nextSegments.push({
-          fill,
-          fillOpacity: isDimmed ? 0.4 : 1,
-          stageId: stage.id,
-          taskId: isExternalStage ? undefined : task.taskId,
-          title: resolveTaskTitle(task),
-          width: 1,
-        });
-        totalCount += 1;
-      });
-      const waitingCount = expectedStageTotal - orderedTasks.length;
-      if (waitingCount > 0) {
-        nextSegments.push({
-          fill: waitingColor,
-          fillOpacity: 1,
-          stageId: stage.id,
-          taskId: undefined,
-          title: `${stage.title ?? stage.id} pending tasks`,
-          width: waitingCount,
-        });
-        totalCount += waitingCount;
-      }
-    });
-    return { segments: nextSegments, stageOffsets: nextStageOffsets, stageCounts: nextStageCounts };
-  }, [
-    failedColor,
-    pausedColor,
-    filter.completedMode,
-    filter.failedMode,
-    filter.skippedMode,
-    resolveTaskTitle,
-    runningColor,
-    skippedColor,
-    stages,
-    successColor,
-    tasksByStage,
-    stageTotals,
-    waitingColor,
-  ]);
-  const viewWidth = Math.max(1, segments.reduce((total, segment) => total + segment.width, 0));
-  const rectHeight = 20;
-  if (viewportStageId && viewportStartIndex !== null && viewportEndIndex !== null) {
-    const stageOffset = stageOffsets.get(viewportStageId);
-    if (stageOffset !== undefined) {
-      viewportStartGlobal = stageOffset + viewportStartIndex;
-      viewportEndGlobal = stageOffset + viewportEndIndex;
-    }
-  }
-  const outRangePaddingRatio = 0.2;
-  const outRangeY = outRangePaddingRatio;
-  const outRangeHeight = 1 - (outRangePaddingRatio * 2);
-  const isSelfActiveStage = Boolean(
-    activeStageId
-    && stages.some((stage) => stage.id === activeStageId)
-  );
-  const showFlowBand = buildStatus === 'running' && isSelfActiveStage;
-  let flowBandRange: { x: number; width: number } | null = null;
-  if (showFlowBand && activeStageId) {
-    const stageOffset = stageOffsets.get(activeStageId);
-    const stageCount = stageCounts.get(activeStageId) ?? 0;
-    if (stageOffset !== undefined && stageCount > 0) {
-      flowBandRange = { x: stageOffset, width: stageCount };
-    }
-  }
-  const flowBandWidth = viewWidth * 0.1;
-  const isDraggingRef = useRef(false);
-  const dragDebounceRef = useRef<ReturnType<typeof setTimeout> | null>(null);
-  const dragDebounceMs = 80;
-
-  const resolveTargetFromPosition = useCallback((position: number) => {
-    if (segments.length === 0) return null;
-    let offset = 0;
-    let index = 0;
-    for (const segment of segments) {
-      const next = offset + segment.width;
-      if (position >= offset && position < next) {
-        return { segment, index };
-      }
-      offset = next;
-      index += 1;
-    }
-    const lastSegment = segments[segments.length - 1];
-    if (!lastSegment) return null;
-    return { segment: lastSegment, index: segments.length - 1 };
-  }, [segments]);
-
-  const resolveNearestInteractiveSegment = useCallback((index: number) => {
-    for (let offset = 0; offset < segments.length; offset += 1) {
-      const left = index - offset;
-      if (left >= 0) {
-        const candidate = segments[left];
-        if (candidate?.taskId) return candidate;
-      }
-      const right = index + offset;
-      if (right < segments.length) {
-        const candidate = segments[right];
-        if (candidate?.taskId) return candidate;
-      }
-    }
-    return null;
-  }, [segments]);
-
-  const updateScrollTargetFromClientX = useCallback((clientX: number, rect: DOMRect) => {
-    if (!rect || rect.width <= 0 || viewWidth <= 0) return;
-    const ratio = (clientX - rect.left) / rect.width;
-    const clampedRatio = Math.max(0, Math.min(1, ratio));
-    const position = clampedRatio * viewWidth;
-    const resolved = resolveTargetFromPosition(position);
-    if (!resolved) return;
-    const segment = resolved.segment.taskId
-      ? resolved.segment
-      : resolveNearestInteractiveSegment(resolved.index);
-    if (!segment?.taskId) return;
-    setScrollTarget({
-      stageId: segment.stageId,
-      taskId: segment.taskId,
-      requestedAt: Date.now(),
-    });
-  }, [resolveNearestInteractiveSegment, resolveTargetFromPosition, setScrollTarget, viewWidth]);
-
-  const handlePointerDown = useCallback((event: React.PointerEvent<SVGSVGElement>) => {
-    event.preventDefault();
-    isDraggingRef.current = true;
-    event.currentTarget.setPointerCapture?.(event.pointerId);
-    updateScrollTargetFromClientX(event.clientX, event.currentTarget.getBoundingClientRect());
-  }, [updateScrollTargetFromClientX]);
-
-  const handlePointerMove = useCallback((event: React.PointerEvent<SVGSVGElement>) => {
-    if (!isDraggingRef.current) return;
-    if (dragDebounceRef.current) {
-      clearTimeout(dragDebounceRef.current);
-    }
-    const rect = event.currentTarget.getBoundingClientRect();
-    const clientX = event.clientX;
-    dragDebounceRef.current = setTimeout(() => {
-      dragDebounceRef.current = null;
-      updateScrollTargetFromClientX(clientX, rect);
-    }, dragDebounceMs);
-  }, [updateScrollTargetFromClientX]);
-
-  const handlePointerUp = useCallback((event: React.PointerEvent<SVGSVGElement>) => {
-    if (!isDraggingRef.current) return;
-    isDraggingRef.current = false;
-    if (dragDebounceRef.current) {
-      clearTimeout(dragDebounceRef.current);
-      dragDebounceRef.current = null;
-    }
-    updateScrollTargetFromClientX(event.clientX, event.currentTarget.getBoundingClientRect());
-    event.currentTarget.releasePointerCapture?.(event.pointerId);
-  }, [updateScrollTargetFromClientX]);
-
-  return (
-    <Box
-      sx={{
-        width: '100%',
-        height: rectHeight,
-        position: 'relative',
-        overflow: 'hidden',
-      }}
-    >
-      <svg
-        width="100%"
-        height={rectHeight}
-        viewBox={`0 0 ${viewWidth} 1`}
-        preserveAspectRatio="none"
-        shapeRendering="crispEdges"
-        onPointerDown={handlePointerDown}
-        onPointerMove={handlePointerMove}
-        onPointerUp={handlePointerUp}
-        onPointerCancel={handlePointerUp}
-        onPointerLeave={handlePointerUp}
-        style={{ touchAction: 'none' }}
-      >
-        <title>{activeStageId}</title>
-        {flowBandRange ? (
-          <defs>
-            <clipPath id={`task-progress-flow-${flowBandClipId}`}>
-              <rect
-                x={flowBandRange.x}
-                y={0}
-                width={flowBandRange.width}
-                height={1}
-              />
-            </clipPath>
-          </defs>
-        ) : null}
-        {segments.length > 0 ? (() => {
-          let offset = 0;
-          return segments.map((segment, index) => {
-            const segmentStart = offset;
-            const x = segmentStart;
-            const segmentEnd = segmentStart + segment.width - 1;
-            const isInViewport =
-              viewportStartGlobal !== null
-              && viewportEndGlobal !== null
-              && segmentStart <= viewportEndGlobal
-              && segmentEnd >= viewportStartGlobal;
-            const y = isInViewport ? 0 : outRangeY;
-            const height = isInViewport ? 1 : outRangeHeight;
-            offset += segment.width;
-            const handleActivate = (event?: React.MouseEvent | React.KeyboardEvent) => {
-              event?.preventDefault();
-              if (!segment.taskId) return;
-              setScrollTarget({
-                stageId: segment.stageId,
-                taskId: segment.taskId,
-                requestedAt: Date.now(),
-              });
-            };
-            const rect = (
-              <rect
-                x={x}
-                y={y}
-                width={segment.width}
-                height={height}
-                fill={segment.fill}
-                fillOpacity={segment.fillOpacity}
-              />
-            );
-            if (!segment.taskId) {
-              return (
-                <g key={`task-${index.toString()}`}>
-                  {rect}
-                  <title>{segment.title ?? 'Task segment'}</title>
-                </g>
-              );
-            }
-            return (
-              <a
-                key={`task-${index.toString()}`}
-                href={`#task-${segment.taskId}`}
-                onClick={handleActivate}
-                onKeyDown={(event) => {
-                  if (event.key === 'Enter' || event.key === ' ') {
-                    handleActivate(event);
-                  }
-                }}
-                aria-label={`Scroll to ${segment.stageId} task`}
-                style={{ cursor: 'pointer' }}
-              >
-                {rect}
-                <title>{segment.title ?? 'Task segment'}</title>
-              </a>
-            );
-          });
-        })() : (
-          <rect
-            key="task-empty"
-            x={0}
-            y={outRangeY}
-            width={1}
-            height={outRangeHeight}
-            fill={emptyStageColor}
-          />
-        )}
-        {showFlowBand && flowBandRange ? (
-          <rect
-            x={-flowBandWidth}
-            y={0}
-            width={flowBandWidth}
-            height={1}
-            fill="#ffffff80"
-            clipPath={`url(#task-progress-flow-${flowBandClipId})`}
-          >
-            <animate
-              attributeName="x"
-              from={-flowBandWidth}
-              to={viewWidth}
-              dur="1.6s"
-              repeatCount="indefinite"
-            />
-          </rect>
-        ) : null}
-      </svg>
-    </Box>
-  );
-};
-
-const BuildProgressStageContent = ({
-  showHeader,
-  stage,
-  stageValue,
-  tasksByStage,
-  paneProgress,
-  isTaskSummaryLoading,
-  isTasksLoading,
-  resolveStatusLabel,
-  resolveStatusColor,
-  resolveTaskTitle,
-  t,
-  matchesSearchQuery,
-}: {
-  showHeader?: boolean;
-  stage: BuildStage;
-  stageValue: number;
-  tasksByStage: Record<string, TaskItemWithMetadata[]>;
-  paneProgress?: Array<{ paneId?: string; progress?: number; taskCount?: number; completedCount?: number; status?: string }>;
-  isTaskSummaryLoading: boolean;
-  isTasksLoading: boolean;
-  resolveStatusLabel: (statusValue?: string, skipped?: boolean) => string;
-  resolveStatusColor: (statusValue?: string, skipped?: boolean) => 'default' | 'success' | 'error' | 'warning' | 'info';
-  resolveTaskTitle: (task: TaskItemWithMetadata) => string;
-  t: (key: string, fallback: string) => string;
-  matchesSearchQuery: (task: TaskItemWithMetadata) => boolean;
-}) => {
-  const filter = useBuildStageFilter();
-  const stageTasks = tasksByStage[stage.id] ?? [];
-  const scrollTarget = useAtomValue(taskScrollTargetAtom);
-  const setScrollTarget = useSetAtom(taskScrollTargetAtom);
-  const viewportRange = useAtomValue(taskViewportRangeAtom);
-  const listScrollRef = useRef<HTMLDivElement | null>(null);
-  const listWrapperRef = useRef<HTMLDivElement | null>(null);
-  const scrollToTaskId = scrollTarget?.stageId === stage.id ? scrollTarget.taskId : undefined;
-  const scrollRequestId = scrollTarget?.requestedAt;
-  const disableVirtualization = typeof window !== 'undefined'
-    && new URLSearchParams(window.location.search).has('noTaskVirtual');
-  const filteredTasks = stageTasks.filter((task) => {
-    if (!matchesSearchQuery(task)) return false;
-    if (isTaskSkipped(task.display, task.message)) return filter.skippedMode;
-    if (task.status === 'failed') return filter.failedMode;
-    if (task.status === 'completed' || task.status === 'recycled') return filter.completedMode;
-    return true;
-  });
-  const orderedTasks = stage.id === 'vt'
-    ? sortVectorTileTasks(filteredTasks)
-    : stage.id === 'transform'
-      ? sortTransformTasks(filteredTasks)
-      : filteredTasks;
-  const displayTasks = orderedTasks;
-  const hasTasks = filteredTasks.length > 0;
-  const stagePane = paneProgress?.find((entry) => entry.paneId === stage.id);
-  const hasSummaryTasks = (stagePane?.taskCount ?? 0) > 0;
-  const showSummarySkeleton = isTaskSummaryLoading && !hasTasks && !hasSummaryTasks;
-  const showTaskSkeleton = !hasTasks && !showSummarySkeleton && isTasksLoading;
-  const requestedTargetIndex = useMemo(() => {
-    if (!scrollToTaskId) return null;
-    const index = orderedTasks.findIndex((task) => task.taskId === scrollToTaskId);
-    return index >= 0 ? index : null;
-  }, [orderedTasks, scrollToTaskId]);
-  const viewportIndices = useMemo(() => {
-    if (orderedTasks.length === 0) return null;
-    if (viewportRange?.stageId !== stage.id || viewportRange == null) return null;
-    const maxIndex = orderedTasks.length - 1;
-    const clampedStart = Math.min(Math.max(viewportRange.startIndex, 0), maxIndex);
-    const clampedEnd = Math.min(Math.max(viewportRange.endIndex, clampedStart), maxIndex);
-    return {
-      startIndex: clampedStart,
-      endIndex: clampedEnd,
-    };
-  }, [orderedTasks.length, stage.id, viewportRange]);
-  const currentIndex = useMemo(() => {
-    if (orderedTasks.length === 0) return null;
-    if (viewportIndices == null) return 0;
-    if (requestedTargetIndex !== null
-      && requestedTargetIndex >= viewportIndices.startIndex
-      && requestedTargetIndex <= viewportIndices.endIndex) {
-      return requestedTargetIndex;
-    }
-    return Math.floor((viewportIndices.startIndex + viewportIndices.endIndex) / 2);
-  }, [orderedTasks.length, requestedTargetIndex, viewportIndices]);
-  const activeTargetIndices = useMemo(() => (
-    orderedTasks.reduce<number[]>((acc, task, index) => {
-      if (task.status === 'running' || task.status === 'queued') {
-        acc.push(index);
-      }
-      return acc;
-    }, [])
-  ), [orderedTasks]);
-  const hasRunningTask = useMemo(
-    () => orderedTasks.some((task) => task.status === 'running'),
-    [orderedTasks],
-  );
-  const hasOnlyQueuedInViewport = useMemo(() => {
-    if (viewportIndices == null) return false;
-    for (let i = viewportIndices.startIndex; i <= viewportIndices.endIndex; i += 1) {
-      if (orderedTasks[i]?.status !== 'queued') return false;
-    }
-    return true;
-  }, [orderedTasks, viewportIndices]);
-  const upTargetIndex = useMemo(() => {
-    if (currentIndex == null) return null;
-    for (let i = activeTargetIndices.length - 1; i >= 0; i -= 1) {
-      const candidate = activeTargetIndices[i];
-      if (candidate == null) continue;
-      if (candidate < currentIndex) return candidate;
-    }
-    return null;
-  }, [activeTargetIndices, currentIndex]);
-  const downTargetIndex = useMemo(() => {
-    if (currentIndex == null) return null;
-    for (const candidate of activeTargetIndices) {
-      if (candidate > currentIndex) return candidate;
-    }
-    return null;
-  }, [activeTargetIndices, currentIndex]);
-  const upTargetTaskId = upTargetIndex === null ? undefined : orderedTasks[upTargetIndex]?.taskId;
-  const downTargetTaskId = downTargetIndex === null ? undefined : orderedTasks[downTargetIndex]?.taskId;
-  const isScrollTargetReached = requestedTargetIndex !== null
-    && currentIndex !== null
-    && currentIndex === requestedTargetIndex;
-  const viewportStartIndex = viewportIndices?.startIndex;
-  const viewportEndIndex = viewportIndices?.endIndex;
-  const showUpArrow = !isScrollTargetReached
-    && !hasOnlyQueuedInViewport
-    && hasRunningTask
-    && upTargetIndex !== null
-    && currentIndex !== null
-    && upTargetIndex < currentIndex
-    && (
-      viewportStartIndex == null
-      || upTargetIndex < viewportStartIndex
-    );
-  const showDownArrow = !isScrollTargetReached
-    && !hasOnlyQueuedInViewport
-    && hasRunningTask
-    && downTargetIndex !== null
-    && currentIndex !== null
-    && currentIndex < downTargetIndex
-    && (
-      viewportEndIndex == null
-      || downTargetIndex > viewportEndIndex
-    );
-  const handleScrollToDirection = useCallback((direction: 'up' | 'down') => {
-    const targetTaskId = direction === 'up' ? upTargetTaskId : downTargetTaskId;
-    if (!targetTaskId) return;
-    setScrollTarget({
-      stageId: stage.id,
-      taskId: targetTaskId,
-      requestedAt: Date.now(),
-    });
-  }, [downTargetTaskId, setScrollTarget, stage.id, upTargetTaskId]);
-  useEffect(() => {
-    const wrapper = listWrapperRef.current;
-    if (!wrapper) return;
-    const handleWheel = (event: globalThis.WheelEvent) => {
-      event.stopPropagation();
-      const scrollEl = listScrollRef.current;
-      if (!scrollEl) return;
-      if (scrollEl.scrollHeight <= scrollEl.clientHeight) {
-        event.preventDefault();
-        return;
-      }
-      scrollEl.scrollTop += event.deltaY;
-      event.preventDefault();
-    };
-    wrapper.addEventListener('wheel', handleWheel, { passive: false });
-    return () => {
-      wrapper.removeEventListener('wheel', handleWheel);
-    };
-  }, []);
-
-  return (
-    <Stack spacing={1} sx={{ p: 2, height: '100%', minHeight: 0 }}>
-      {showSummarySkeleton ? (
-        <>
-          <Skeleton variant="text" width="40%" />
-          <Skeleton variant="text" width="70%" />
-          <Skeleton variant="rounded" height={88} />
-        </>
-      ) : showTaskSkeleton ? (
-        <>
-          <Skeleton variant="text" width="35%" />
-          <Skeleton variant="rounded" height={88} />
-        </>
-      ) : !hasTasks ? (
-        <>
-          {showHeader ? (
-            <>
-              <Typography variant="subtitle2">{stage.title}</Typography>
-              {stage.description ? (
-                <Typography variant="body2" color="text.secondary">
-                  {stage.description}
-                </Typography>
-              ) : null}
-            </>
-          ) : null}
-          <Typography variant="caption" color="text.secondary">
-            {hasSummaryTasks
-              ? t('stage.tasks.summaryOnly', 'Tasks are summarized. Detailed list is unavailable.')
-              : t('stage.tasks.empty', 'No tasks yet.')}
-          </Typography>
-        </>
-      ) : (
-        <Box
-          sx={{
-            position: 'relative',
-            flex: 1,
-            minHeight: 0,
-            height: '100%',
-            display: 'flex',
-            flexDirection: 'column',
-          }}
-          ref={listWrapperRef}
-        >
-          <TaskItemCardListCard
-            ref={listScrollRef}
-            stageId={stage.id}
-            tasks={displayTasks}
-            stageValue={stageValue}
-            resolveStatusLabel={resolveStatusLabel}
-            resolveStatusColor={resolveStatusColor}
-            resolveTaskTitle={resolveTaskTitle}
-            scrollToTaskId={scrollToTaskId}
-            scrollRequestId={scrollRequestId}
-            virtualize={!disableVirtualization}
-          />
-          {showUpArrow ? (
-            <Tooltip title={t('stage.progress.scrollToRunningUp', 'Scroll up to running or queued task')}>
-              <IconButton
-                aria-label={t('stage.progress.scrollToRunningUp', 'Scroll up to running or queued task')}
-                color="primary"
-                onClick={() => handleScrollToDirection('up')}
-                sx={{
-                  position: 'absolute',
-                  left: '50%',
-                  top: 0,
-                  transform: 'translateX(-50%)',
-                  bgcolor: 'transparent',
-                  boxShadow: 'none',
-                  zIndex: 2,
-                  width: 56,
-                  height: 56,
-                  '&:hover': { bgcolor: 'transparent' },
-                }}
-              >
-                <ArrowCircleUpIcon sx={{ fontSize: 48 }} />
-              </IconButton>
-            </Tooltip>
-          ) : null}
-          {showDownArrow ? (
-            <Tooltip title={t('stage.progress.scrollToRunningDown', 'Scroll down to running or queued task')}>
-              <IconButton
-                aria-label={t('stage.progress.scrollToRunningDown', 'Scroll down to running or queued task')}
-                color="primary"
-                onClick={() => handleScrollToDirection('down')}
-                sx={{
-                  position: 'absolute',
-                  left: '50%',
-                  bottom: 0,
-                  transform: 'translateX(-50%)',
-                  bgcolor: 'transparent',
-                  boxShadow: 'none',
-                  zIndex: 2,
-                  width: 56,
-                  height: 56,
-                  '&:hover': { bgcolor: 'transparent' },
-                }}
-              >
-                <ArrowCircleDownIcon sx={{ fontSize: 48 }} />
-              </IconButton>
-            </Tooltip>
-          ) : null}
-        </Box>
-      )}
-    </Stack>
-  );
-};
+} from '../../../../common/types/index.js';
+import { BuildProgressStageContent } from './BuildProgressStageContent.js';
+import { useShapeBuildProgressPanel } from '../useShapeBuildProgressPanel.js';
+import { TaskProgressBar } from './TaskProgressBar.js';
+import { useShapeBuildCacheActions } from '../../../hooks/useShapeBuildCacheActions.js';
 
 export const ShapeBuildProgressPanel = ({
   data,
@@ -1239,7 +581,7 @@ export const ShapeBuildProgressPanel = ({
   ]);
 
   const stageHeaderMeta = useMemo(() => (
-    stages.reduce<Record<string, ReactNode>>((acc, stage) => {
+    stages.reduce<Record<string, JSX.Element>>((acc, stage) => {
       acc[stage.id] = buildTimingSummary(stage.id);
       return acc;
     }, {})
