+++ plugins/shape-plugin/src/ui/components/build-progress/ShapeBuildProgressPanel/TaskProgressBar.tsx
@@ -1,0 +1,405 @@
+import { useCallback, useId, useMemo, useRef } from 'react';
+import { Box } from '@mui/material';
+import { useTheme } from '@mui/material/styles';
+import { useAtomValue, useSetAtom } from 'jotai';
+import { Skeleton } from '@mui/material';
+import { useBuildStageFilter } from '@hierarchidb/components';
+import type { BuildStage } from '@hierarchidb/components/build-stage';
+import { isTaskSkipped } from '../../../../common/utils/taskMessages.ts';
+import { taskScrollTargetAtom, taskViewportRangeAtom } from '../../../atoms/shapeBuildProgressAtoms.js';
+import type { TaskItemWithMetadata } from '../TaskItemCardListCard/TaskItemCardListCard.tsx';
+import type { TaskProgressSummary } from '../../../atoms/shapeBuildProgressAtoms.js';
+import { sortTransformTasks, sortVectorTileTasks } from '../useTaskItemCardList.ts';
+
+type TaskProgressBarProps = {
+  stages: BuildStage[];
+  tasksByStage: Record<string, TaskItemWithMetadata[]>;
+  stageTotals?: TaskProgressSummary['stageTotals'];
+  buildStatus: TaskProgressSummary['buildStatus'];
+  activeStageId?: string | null;
+  resolveTaskTitle: (task: TaskItemWithMetadata) => string;
+};
+
+type Segment = {
+  fill: string;
+  fillOpacity: number;
+  stageId: string;
+  taskId?: string;
+  title: string;
+  width: number;
+};
+
+export const TaskProgressBar = ({
+  stages,
+  tasksByStage,
+  stageTotals,
+  buildStatus,
+  activeStageId,
+  resolveTaskTitle,
+}: TaskProgressBarProps) => {
+  const theme = useTheme();
+  const filter = useBuildStageFilter();
+  const flowBandClipId = useId().replace(/:/g, '');
+  const setScrollTarget = useSetAtom(taskScrollTargetAtom);
+  const viewportRange = useAtomValue(taskViewportRangeAtom);
+  const waitingColor = theme.palette.grey[300];
+  const emptyStageColor = buildStatus === 'failed' ? theme.palette.error.main : theme.palette.grey[500];
+  const runningColor = theme.palette.info.main;
+  const failedColor = theme.palette.error.main;
+  const skippedColor = theme.palette.warning.main;
+  const viewportStageId = viewportRange?.stageId;
+  const viewportStartId = viewportRange?.startTaskId;
+  const viewportEndId = viewportRange?.endTaskId;
+  let viewportStartGlobal: number | null = null;
+  let viewportEndGlobal: number | null = null;
+  let viewportStartIndex: number | null = null;
+  let viewportEndIndex: number | null = null;
+
+  if (viewportStageId && viewportStartId && viewportEndId) {
+    const stageTasks = tasksByStage[viewportStageId] ?? [];
+    const orderedStageTasks = viewportStageId === 'vt'
+      ? sortVectorTileTasks(stageTasks)
+      : viewportStageId === 'transform'
+        ? sortTransformTasks(stageTasks)
+      : stageTasks;
+    const startIndex = orderedStageTasks.findIndex((task) => task.taskId === viewportStartId);
+    const endIndex = orderedStageTasks.findIndex((task) => task.taskId === viewportEndId);
+    if (startIndex >= 0 && endIndex >= 0) {
+      viewportStartIndex = Math.min(startIndex, endIndex);
+      viewportEndIndex = Math.max(startIndex, endIndex);
+    }
+  }
+
+  const successColor = theme.palette.success.main;
+  const pausedColor = theme.palette.warning.main;
+  const { segments, stageOffsets, stageCounts } = useMemo(() => {
+    const nextSegments: Segment[] = [];
+    const nextStageOffsets = new Map<string, number>();
+    const nextStageCounts = new Map<string, number>();
+    let totalCount = 0;
+
+    stages.forEach((stage) => {
+      const fallbackStageId = stage.id === 'transform'
+        && stages.length === 1
+        && (tasksByStage.transform?.length ?? 0) === 0
+        && (tasksByStage.fetch?.length ?? 0) > 0
+        ? 'fetch'
+        : stage.id;
+      const sourceStageId = fallbackStageId;
+      const stageTasks = tasksByStage[sourceStageId] ?? [];
+      nextStageOffsets.set(stage.id, totalCount);
+      const plannedStageTotal = Math.max(0, stageTotals?.[stage.id]?.total ?? 0);
+      const orderedTasks = stage.id === 'vt'
+        ? sortVectorTileTasks(stageTasks)
+        : stage.id === 'transform'
+          ? sortTransformTasks(stageTasks)
+        : stageTasks;
+      const expectedStageTotal = Math.max(orderedTasks.length, plannedStageTotal);
+      nextStageCounts.set(stage.id, expectedStageTotal);
+
+      orderedTasks.forEach((task) => {
+        const statusValue = (task.status ?? '').toString().toLowerCase();
+        let fill = waitingColor;
+        const isSkipped = isTaskSkipped(task.display, task.message);
+        if (isSkipped) {
+          fill = skippedColor;
+        } else if (statusValue === 'completed' || statusValue === 'recycled') {
+          fill = successColor;
+        } else if (statusValue === 'failed') {
+          fill = failedColor;
+        } else if (statusValue === 'running') {
+          fill = runningColor;
+        } else if (statusValue === 'paused') {
+          fill = pausedColor;
+        }
+        const isDimmed =
+          (isSkipped && !filter.skippedMode)
+          || (statusValue === 'failed' && !filter.failedMode)
+          || ((statusValue === 'completed' || statusValue === 'recycled') && !filter.completedMode);
+        const isExternalStage = sourceStageId !== stage.id;
+        nextSegments.push({
+          fill,
+          fillOpacity: isDimmed ? 0.4 : 1,
+          stageId: stage.id,
+          taskId: isExternalStage ? undefined : task.taskId,
+          title: resolveTaskTitle(task),
+          width: 1,
+        });
+        totalCount += 1;
+      });
+
+      const waitingCount = expectedStageTotal - orderedTasks.length;
+      if (waitingCount > 0) {
+        nextSegments.push({
+          fill: waitingColor,
+          fillOpacity: 1,
+          stageId: stage.id,
+          taskId: undefined,
+          title: `${stage.title ?? stage.id} pending tasks`,
+          width: waitingCount,
+        });
+        totalCount += waitingCount;
+      }
+    });
+
+    return { segments: nextSegments, stageOffsets: nextStageOffsets, stageCounts: nextStageCounts };
+  }, [
+    filter.completedMode,
+    filter.failedMode,
+    filter.skippedMode,
+    stageTotals,
+    stages,
+    tasksByStage,
+    waitingColor,
+    skippedColor,
+    runningColor,
+    failedColor,
+    pausedColor,
+    successColor,
+    resolveTaskTitle,
+  ]);
+
+  const viewWidth = Math.max(1, segments.reduce((total, segment) => total + segment.width, 0));
+  const rectHeight = 20;
+  if (viewportStageId && viewportStartIndex !== null && viewportEndIndex !== null) {
+    const stageOffset = stageOffsets.get(viewportStageId);
+    if (stageOffset !== undefined) {
+      viewportStartGlobal = stageOffset + viewportStartIndex;
+      viewportEndGlobal = stageOffset + viewportEndIndex;
+    }
+  }
+
+  const outRangePaddingRatio = 0.2;
+  const outRangeY = outRangePaddingRatio;
+  const outRangeHeight = 1 - (outRangePaddingRatio * 2);
+  const isSelfActiveStage = Boolean(
+    activeStageId
+    && stages.some((stage) => stage.id === activeStageId)
+  );
+  const showFlowBand = buildStatus === 'running' && isSelfActiveStage;
+  const flowBandRange = (() => {
+    if (!showFlowBand || !activeStageId) return null;
+    const stageOffset = stageOffsets.get(activeStageId);
+    const stageCount = stageCounts.get(activeStageId) ?? 0;
+    if (stageOffset !== undefined && stageCount > 0) {
+      return { x: stageOffset, width: stageCount };
+    }
+    return null;
+  })();
+
+  const flowBandWidth = viewWidth * 0.1;
+  const isDraggingRef = useRef(false);
+  const dragDebounceRef = useRef<ReturnType<typeof setTimeout> | null>(null);
+  const dragDebounceMs = 80;
+
+  const resolveTargetFromPosition = useCallback((position: number) => {
+    if (segments.length === 0) return null;
+    let offset = 0;
+    let index = 0;
+    for (const segment of segments) {
+      const next = offset + segment.width;
+      if (position >= offset && position < next) {
+        return { segment, index };
+      }
+      offset = next;
+      index += 1;
+    }
+    const lastSegment = segments[segments.length - 1];
+    if (!lastSegment) return null;
+    return { segment: lastSegment, index: segments.length - 1 };
+  }, [segments]);
+
+  const resolveNearestInteractiveSegment = useCallback((index: number) => {
+    for (let offset = 0; offset < segments.length; offset += 1) {
+      const left = index - offset;
+      if (left >= 0) {
+        const candidate = segments[left];
+        if (candidate?.taskId) return candidate;
+      }
+      const right = index + offset;
+      if (right < segments.length) {
+        const candidate = segments[right];
+        if (candidate?.taskId) return candidate;
+      }
+    }
+    return null;
+  }, [segments]);
+
+  const updateScrollTargetFromClientX = useCallback((clientX: number, rect: DOMRect) => {
+    if (!rect || rect.width <= 0 || viewWidth <= 0) return;
+    const ratio = (clientX - rect.left) / rect.width;
+    const clampedRatio = Math.max(0, Math.min(1, ratio));
+    const position = clampedRatio * viewWidth;
+    const resolved = resolveTargetFromPosition(position);
+    if (!resolved) return;
+    const segment = resolved.segment.taskId
+      ? resolved.segment
+      : resolveNearestInteractiveSegment(resolved.index);
+    if (!segment?.taskId) return;
+    setScrollTarget({
+      stageId: segment.stageId,
+      taskId: segment.taskId,
+      requestedAt: Date.now(),
+    });
+  }, [resolveNearestInteractiveSegment, resolveTargetFromPosition, setScrollTarget, viewWidth]);
+
+  const handlePointerDown = useCallback((event: React.PointerEvent<SVGSVGElement>) => {
+    event.preventDefault();
+    isDraggingRef.current = true;
+    event.currentTarget.setPointerCapture?.(event.pointerId);
+    updateScrollTargetFromClientX(event.clientX, event.currentTarget.getBoundingClientRect());
+  }, [updateScrollTargetFromClientX]);
+
+  const handlePointerMove = useCallback((event: React.PointerEvent<SVGSVGElement>) => {
+    if (!isDraggingRef.current) return;
+    if (dragDebounceRef.current) {
+      clearTimeout(dragDebounceRef.current);
+    }
+    const rect = event.currentTarget.getBoundingClientRect();
+    const clientX = event.clientX;
+    dragDebounceRef.current = setTimeout(() => {
+      dragDebounceRef.current = null;
+      updateScrollTargetFromClientX(clientX, rect);
+    }, dragDebounceMs);
+  }, [updateScrollTargetFromClientX]);
+
+  const handlePointerUp = useCallback((event: React.PointerEvent<SVGSVGElement>) => {
+    if (!isDraggingRef.current) return;
+    isDraggingRef.current = false;
+    if (dragDebounceRef.current) {
+      clearTimeout(dragDebounceRef.current);
+      dragDebounceRef.current = null;
+    }
+    updateScrollTargetFromClientX(event.clientX, event.currentTarget.getBoundingClientRect());
+    event.currentTarget.releasePointerCapture?.(event.pointerId);
+  }, [updateScrollTargetFromClientX]);
+
+  return (
+    <Box
+      sx={{
+        width: '100%',
+        height: rectHeight,
+        position: 'relative',
+        overflow: 'hidden',
+      }}
+    >
+      <svg
+        width="100%"
+        height={rectHeight}
+        viewBox={`0 0 ${viewWidth} 1`}
+        preserveAspectRatio="none"
+        shapeRendering="crispEdges"
+        onPointerDown={handlePointerDown}
+        onPointerMove={handlePointerMove}
+        onPointerUp={handlePointerUp}
+        onPointerCancel={handlePointerUp}
+        onPointerLeave={handlePointerUp}
+        style={{ touchAction: 'none' }}
+      >
+        <title>{activeStageId}</title>
+        {flowBandRange ? (
+          <defs>
+            <clipPath id={`task-progress-flow-${flowBandClipId}`}>
+              <rect
+                x={flowBandRange.x}
+                y={0}
+                width={flowBandRange.width}
+                height={1}
+              />
+            </clipPath>
+          </defs>
+        ) : null}
+        {segments.length > 0 ? (() => {
+          let offset = 0;
+          return segments.map((segment, index) => {
+            const segmentStart = offset;
+            const x = segmentStart;
+            const segmentEnd = segmentStart + segment.width - 1;
+            const isInViewport =
+              viewportStartGlobal !== null
+              && viewportEndGlobal !== null
+              && segmentStart <= viewportEndGlobal
+              && segmentEnd >= viewportStartGlobal;
+            const y = isInViewport ? 0 : outRangeY;
+            const height = isInViewport ? 1 : outRangeHeight;
+            offset += segment.width;
+            const handleActivate = (event?: React.MouseEvent | React.KeyboardEvent) => {
+              event?.preventDefault();
+              if (!segment.taskId) return;
+              setScrollTarget({
+                stageId: segment.stageId,
+                taskId: segment.taskId,
+                requestedAt: Date.now(),
+              });
+            };
+            const rect = (
+              <rect
+                x={x}
+                y={y}
+                width={segment.width}
+                height={height}
+                fill={segment.fill}
+                fillOpacity={segment.fillOpacity}
+              />
+            );
+            if (!segment.taskId) {
+              return (
+                <g key={`task-${index.toString()}`}>
+                  {rect}
+                  <title>{segment.title ?? 'Task segment'}</title>
+                </g>
+              );
+            }
+            return (
+              <a
+                key={`task-${index.toString()}`}
+                href={`#task-${segment.taskId}`}
+                onClick={handleActivate}
+                onKeyDown={(event) => {
+                  if (event.key === 'Enter' || event.key === ' ') {
+                    handleActivate(event);
+                  }
+                }}
+                aria-label={`Scroll to ${segment.stageId} task`}
+                style={{ cursor: 'pointer' }}
+              >
+                {rect}
+                <title>{segment.title ?? 'Task segment'}</title>
+              </a>
+            );
+          });
+        })() : (
+          <rect
+            key="task-empty"
+            x={0}
+            y={outRangeY}
+            width={1}
+            height={outRangeHeight}
+            fill={emptyStageColor}
+          />
+        )}
+        {showFlowBand && flowBandRange ? (
+          <rect
+            x={-flowBandWidth}
+            y={0}
+            width={flowBandWidth}
+            height={1}
+            fill="#ffffff80"
+            clipPath={`url(#task-progress-flow-${flowBandClipId})`}
+          >
+            <animate
+              attributeName="x"
+              from={-flowBandWidth}
+              to={viewWidth}
+              dur="1.6s"
+              repeatCount="indefinite"
+            />
+          </rect>
+        ) : null}
+      </svg>
+    </Box>
+  );
+};
+
+export const skeletonTaskProgressBar = <Skeleton variant="rounded" height={20} />;
+export type { Segment as TaskProgressSegment };
