+++ plugins/shape-plugin/src/ui/components/build-progress/TaskItem/TaskItem.tsx
