+++ plugins/shape-plugin/src/ui/components/build-progress/internal/useShapeBuildStepHelpers.ts
@@ -1,0 +1,38 @@
+export { PAUSE_COMMAND_TIMEOUT_MS, SHAPE_NODE_TYPE, UI_POLL_INTERVAL_MS, UI_QUIET_THRESHOLD_MS } from './useShapeBuildStepHelpers/constants.ts';
+export {
+  type ShapeProgressStepDebugConfig,
+  type ShapeProgressStepTracePayload,
+  emitShapeProgressStepTrace,
+  isShapeProgressStepDebugEnabled,
+  readShapeProgressStepDebugConfig,
+} from './useShapeBuildStepHelpers/debug.ts';
+export {
+  hasPositiveElapsed,
+  mergeElapsedByStage,
+  runWithTimeout,
+  shallowEqualNumberRecord,
+  shouldResetElapsedState,
+  sumNumberRecord,
+} from './useShapeBuildStepHelpers/elapsed.ts';
+export {
+  type BuildStartupTransitionWarnStep,
+  resolveDisplayBuildStatus,
+  shouldRefreshTasksSnapshot,
+  toBuildStatus,
+  toProcessingStatus,
+  summarizeSelectionStateFromConfig,
+} from './useShapeBuildStepHelpers/status.ts';
+export {
+  getErrorMessage,
+  resolveBuildSessionRecordForPersistence,
+  summarizeSelectedEntries,
+  toTransitionErrorMessage,
+} from './useShapeBuildStepHelpers/errors.ts';
+export {
+  normalizeStageKey,
+  resolveMostAdvancedInFlightStageId,
+  resolveMostAdvancedRunningStageId,
+  resolveMostAdvancedStageIdByStatus,
+  type StageLikeRunningTask,
+  type StageLikeTask,
+} from './useShapeBuildStepHelpers/stage.ts';
